First run at the problem
The use of two vectors turned out to be a disadvantage. I had imagined the two vectors would simplify printing to the monitor and mathematical operations. 

This exercise taught me  I don't actually remember as much as I thought I did about basic math operations. One vector would have simplified things, and a variable could have been used to store the decimal position. I will fix this in attempt 2. This thinking can be seen in my final operator overload: *. At this point, it became clear using two vectors (one for the decimal number and the other for the whole number) was in fact a disadvantage.